+++ xUnitRevitUtils2021/xru.cs
@@ -18,12 +18,12 @@ namespace xUnitRevitUtils
   {
     public static UIApplication Uiapp { get; set; }
 
-    private static List<Action> Queue { get; set; }
+    private static IList<Action> Queue { get; set; }
     private static ExternalEvent EventHandler { get; set; }
 
     public static SynchronizationContext UiContext { get; set; }
 
-    public static void Initialize(UIApplication uiapp, SynchronizationContext uiContext, ExternalEvent eventHandler, List<Action> queue)
+    public static void Initialize(UIApplication uiapp, SynchronizationContext uiContext, ExternalEvent eventHandler, IList<Action> queue)
     {
       Uiapp = uiapp;
       UiContext = uiContext;
@@ -33,12 +33,11 @@ namespace xUnitRevitUtils
 
     #region utility methods
 
-
     /// <summary>
     /// Returns the selected elements in the active document
     /// </summary>
     /// <returns></returns>
-    public static List<Element> GetActiveSelection()
+    public static IList<Element> GetActiveSelection()
     {
       Assert.NotNull(Uiapp);
 
@@ -59,7 +58,6 @@ namespace xUnitRevitUtils
       Assert.NotNull(doc);
       return doc;
     }
-
 
     /// <summary>
     /// Creates a new empty document
@@ -97,7 +95,6 @@ namespace xUnitRevitUtils
       Assert.NotNull(doc);
       return doc;
     }
-
 
     /// <summary>
     /// Runs an Action in a Revit transaction, uses TaskCompletionSource to communicate when done
@@ -139,7 +136,6 @@ namespace xUnitRevitUtils
       EventHandler.Raise();
 
       return tcs.Task;
-
     }
 
     /// <summary>
@@ -167,7 +163,6 @@ namespace xUnitRevitUtils
       EventHandler.Raise();
 
       return tcs.Task;
-
     }
 
     /// <summary>
@@ -187,7 +182,6 @@ namespace xUnitRevitUtils
         return FailureProcessingResult.Continue;
       }
     }
-
 
     #endregion
   }
